requirements.txt
asgiref==3.2.7
astroid==2.4.1
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
env-file==2020.5.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
public==2019.4.13
pylint==2.5.2
python-http-client==3.2.7
pytz==2020.1
sendgrid==6.3.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
values==2019.4.13
wrapt==1.12.1
